requirements.txt

Django==3.0.4
django-cors-header==2.4.0
djangorestframework=3.12.2
pytz==2020.4
django-heroku==0.3.1
gunicorn==20.0.4
Pillow==7.0.0
whitenoise==5.0.1

